roll

+++ cogs/utils.py
@@ -107,7 +107,8 @@ class Utils:
     async def _help(self, ctx, *, command: str = None):
         # return
         """Shows help about a command or the bot"""
-        await ctx.message.delete()
+        if not isinstance(ctx.channel, discord.DMChannel):
+            await ctx.message.delete()
         try:
             if command is None:
                 p = await HelpPaginator.from_bot(ctx)
